@@ -667,27 +671,30 @@
 		"@types/fs-extra": "^9.0.13",
 		"@types/glob": "^7.2.0",
 		"@types/jest": "^27.4.1",
-		"@types/node": "^17.0.21",
+		"@types/node": "^17.0.23",
 		"@types/node-fetch": "^2.6.1",
 		"@types/power-assert": "^1.5.8",
 		"@types/uuid": "^8.3.4",
-		"@types/vscode": "^1.65.0",
+		"@types/vscode": "^1.66.0",
 		"@types/vscode-webview": "^1.57.0",
-		"@typescript-eslint/eslint-plugin": "^5.14.0",
-		"@typescript-eslint/parser": "^5.14.0",
+		"@typescript-eslint/eslint-plugin": "^5.17.0",
+		"@typescript-eslint/parser": "^5.17.0",
+		"@vitejs/plugin-vue": "^2.3.1",
 		"@vscode/test-electron": "^2.1.3",
-		"electron": "^17.1.2",
-		"esbuild": "^0.14.25",
+		"@vue/eslint-config-typescript": "^10.0.0",
+		"electron": "^17.3.1",
+		"esbuild": "^0.14.29",
 		"esbuild-jest": "^0.5.0",
-		"esbuild-plugin-d.ts": "^1.1.0",
 		"esbuild-register": "^3.3.2",
-		"eslint": "^8.11.0",
+		"eslint": "^8.12.0",
 		"file-loader": "^6.2.0",
 		"glob": "^7.2.0",
 		"jest": "^27.5.1",
 		"rimraf": "^3.0.2",
 		"ts-loader": "^9.2.8",
-		"typescript": "^4.6.2",
+		"typescript": "^4.6.3",
+		"vite": "^2.9.1",
+		"vue-loader": "^17.0.0",
 		"webpack": "^5.70.0",
 		"webpack-cli": "^4.9.2"
 	},
